requirements.txt
Flask==2.3.5
flask-cors==3.1.1
faiss-cpu==1.7.4
sentence-transformers==2.2.2
numpy==1.26.0
google-generativeai==0.1.2
firebase-admin==6.1.1
requests==2.31.0
torch==2.1.0
